Rebuilt CeaDatasetGenerator.py
Rebuilt CeaDatasetGenerator.py. In the first version, the input data range was pregenerated; this simplified encapsulation of the functions but led to an issue whereby some data might lie outside of the scope of CEA to generate results. The newer version corrects for this by testing (using regular expressions) for CEA validity to ensure the viability of the output data.